requirements.txt
strava-cli
pyyaml
